Release

+++ celos-ci/src/main/java/com/collective/celos/ci/mode/TestTask.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2015 Collective, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied.  See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
 package com.collective.celos.ci.mode;
 
 import java.io.File;
@@ -15,6 +30,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 
+import com.collective.celos.Util;
 import org.apache.commons.io.FileUtils;
 import org.apache.log4j.FileAppender;
 import org.apache.log4j.Level;
@@ -22,7 +38,7 @@ import org.apache.log4j.Logger;
 import org.apache.log4j.PatternLayout;
 
 import com.collective.celos.ci.CelosCi;
-import com.collective.celos.ci.config.CelosCiCommandLine;
+import com.collective.celos.ci.config.CiCommandLine;
 import com.collective.celos.ci.config.deploy.CelosCiTarget;
 import com.collective.celos.ci.config.deploy.CelosCiTargetParser;
 import com.collective.celos.ci.mode.test.TestCase;
@@ -37,19 +53,18 @@ public class TestTask extends CelosCi {
 
     private static final String CELOS_CI_DIR = ".celos-ci";
     private static final String TEST_CONFIG_JS_FILE = "test.js";
-    private static final String CELOS_LOG_FILE = "celos.log";
 
     private final List<TestRun> testRuns = Lists.newArrayList();
     private final File celosCiTempDir;
 
-    public TestTask(CelosCiCommandLine commandLine) throws Exception {
+    public TestTask(CiCommandLine commandLine) throws Exception {
         this(commandLine, getConfigJSFile(commandLine), getTempDir());
     }
 
-    TestTask(CelosCiCommandLine commandLine, File configJSFile, File tempDir) throws Exception {
+    TestTask(CiCommandLine commandLine, File configJSFile, File tempDir) throws Exception {
 
         this.celosCiTempDir = tempDir;
-        substituteLoggers();
+        Util.setupLogging(tempDir);
 
         CelosCiTargetParser parser = new CelosCiTargetParser(commandLine.getUserName());
         CelosCiTarget target = parser.parse(commandLine.getTargetUri());
@@ -62,7 +77,7 @@ public class TestTask extends CelosCi {
         }
     }
 
-    private static File getConfigJSFile(CelosCiCommandLine commandLine) {
+    private static File getConfigJSFile(CiCommandLine commandLine) {
         return new File(commandLine.getTestCasesDir(), TEST_CONFIG_JS_FILE);
     }
 
@@ -76,29 +91,13 @@ public class TestTask extends CelosCi {
         return PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString("rwxr-xr-x"));
     }
 
-    private void substituteLoggers() throws IOException {
-        //don't load configuration from log4j.xml -- we're doing this programmatically
-        System.getProperties().setProperty("log4j.defaultInitOverride", "true");
-
-        FileAppender fileAppender = new FileAppender();
-        fileAppender.setFile(new File(celosCiTempDir, CELOS_LOG_FILE).getAbsolutePath());
-
-        PatternLayout patternLayout = new PatternLayout();
-        patternLayout.setConversionPattern("[%d{YYYY-MM-dd HH:mm:ss.SSS}] [%t] %-5p: %m%n");
-        fileAppender.setLayout(patternLayout);
-        fileAppender.activateOptions();
-        Logger.getRootLogger().addAppender(fileAppender);
-        Logger.getRootLogger().setLevel(Level.INFO);
-    }
-
-
     @Override
     public void start() throws Throwable {
         try {
             List<Future> futures = submitTestRuns();
             waitForCompletion(futures);
         } finally {
-            if (allTestCasesDeletedTheirData()) {
+            if (celosCiTempDir.exists() && allTestCasesDeletedTheirData()) {
                 FileUtils.forceDelete(celosCiTempDir);
             }
         }
